feat: implement financial service and money repository
- Added FinancialService to manage balance, earnings, and print quote generation.
- Introduced moneyRepository for balance management and earnings tracking.
- Updated session handling in financial service to support print quote generation.
- Refactored existing services (settlement, serial) to utilize moneyRepository.
- Created state and settings repositories for better state management.
- Added default state schema and normalization logic for consistent state handling.
- Implemented various types for state management, including admin settings and feedback.
- Enhanced lockout utility to use state repository for admin lockout management.
