+++ src/bootstrap/index.ts
@@ -1,0 +1,3 @@
+export { createKioskAppRuntime, type KioskAppRuntime } from './app';
+export { getLocalIPv4 } from './network';
+export { initializeInfrastructure } from './startup';
