+++ src/controllers/index.ts
@@ -1,0 +1,1 @@
+export { registerSystemApiRoutes } from './system-api-controller';
