+++ src/domain/index.ts
@@ -1,0 +1,1 @@
+export { BACKEND_BOUNDARIES, type BackendBoundary, type BackendLayer } from './boundaries';
