+++ src/runtime/guardrails.ts
@@ -1,0 +1,11 @@
+export interface RefactorGuardrails {
+  preserveEndpointContracts: boolean;
+  preserveDbJsonCompatibility: boolean;
+  incrementalMigrationOnly: boolean;
+}
+
+export const BACKEND_REFACTOR_GUARDRAILS: RefactorGuardrails = Object.freeze({
+  preserveEndpointContracts: true,
+  preserveDbJsonCompatibility: true,
+  incrementalMigrationOnly: true,
+});
