+++ src/services/serial.ts
@@ -1,6 +1,5 @@
 import { SerialPort } from 'serialport';
 import { ReadlineParser } from '@serialport/parser-readline';
-import { db } from './db';
 import { Server } from 'socket.io';
 import { adminService } from './admin';
 import {
@@ -15,6 +14,7 @@ import {
 // so we can refuse coin credits when the printer is offline or faulted.
 import { getPrinterTelemetry } from './printer-status';
 import { BLOCKED_STATUSES } from '@/utils';
+import { moneyRepository } from '@/state/repositories';
 
 const ACCEPTED_COINS = new Set([1, 5, 10, 20]);
 const FRAGMENT_WINDOW_MS = 140;
@@ -376,23 +376,23 @@ async function attemptSerialConnection(io: Server, attempt: number) {
       };
 
       const persistBalance = async (coinValue: number) => {
-        db.data!.balance += coinValue;
+        const newBalance = await moneyRepository.incrementBalance(coinValue);
         await adminService.incrementCoinStats(coinValue);
         await adminService.appendAdminLog(
           'coin_accepted',
           `Accepted coin: ${coinValue}`,
           {
             coinValue,
-            balance: db.data!.balance,
+            balance: newBalance,
           },
         );
         console.log(
-          `[SERIAL] ✓ Coin accepted: ₱${coinValue} → new balance: ₱${db.data!.balance}`,
+          `[SERIAL] ✓ Coin accepted: ₱${coinValue} → new balance: ₱${newBalance}`,
         );
-        io.emit('balance', db.data!.balance);
+        io.emit('balance', newBalance);
         io.emit('coinAccepted', {
           value: coinValue,
-          balance: db.data!.balance,
+          balance: newBalance,
         });
       };
