+++ src/state/types/admin-settings.ts
@@ -1,0 +1,18 @@
+export type AdminLockout = {
+  failedAttempts: number;
+  lockedUntil: string | null;
+};
+
+export interface PricingSettings {
+  printPerPage: number;
+  copyPerPage: number;
+  scanDocument: number;
+  colorSurcharge: number;
+}
+
+export interface AdminSettings {
+  pricing: PricingSettings;
+  idleTimeoutSeconds: number;
+  adminPin: string;
+  adminLocalOnly: boolean;
+}
