+++ src/state/types/common.ts
@@ -1,0 +1,1 @@
+export type LogMeta = Record<string, string | number | boolean | null>;
