+++ src/state/types/feedback.ts
@@ -1,0 +1,34 @@
+import type { LogMeta } from './common';
+
+export type FeedbackCategory =
+  | 'service'
+  | 'hardware'
+  | 'software'
+  | 'print'
+  | 'scan'
+  | 'copy'
+  | 'payment'
+  | 'other';
+
+export type FeedbackStatus = 'open' | 'resolved';
+
+export interface FeedbackEntry {
+  id: string;
+  sessionId: string;
+  timestamp: string;
+  comment: string;
+  category: FeedbackCategory | null;
+  rating: number | null;
+  status: FeedbackStatus;
+  resolvedAt?: string | null;
+  meta?: LogMeta;
+}
+
+export interface FeedbackSessionEntry {
+  id: string;
+  token: string;
+  feedbackUrl: string;
+  createdAt: string;
+  expiresAt: string;
+  submittedAt: string | null;
+}
