+++ src/state/types/print.ts
@@ -1,0 +1,2 @@
+export type PrintMode = 'print' | 'copy' | 'scan';
+export type ColorMode = 'colored' | 'grayscale';
