+++ src/state/types/report-issue.ts
@@ -1,0 +1,48 @@
+import type { LogMeta } from './common';
+
+export type ReportIssueCategory =
+  | 'hardware'
+  | 'software'
+  | 'print'
+  | 'copy'
+  | 'scan'
+  | 'payment'
+  | 'network'
+  | 'other';
+
+export type ReportIssueStatus = 'open' | 'acknowledged' | 'resolved';
+
+export interface ReportIssueSessionEntry {
+  id: string;
+  token: string;
+  reportUrl: string;
+  createdAt: string;
+  expiresAt: string;
+  submittedAt: string | null;
+}
+
+export interface ReportIssueAttachmentEntry {
+  id: string;
+  sessionId: string;
+  reportIssueId: string | null;
+  timestamp: string;
+  originalName: string;
+  storedName: string;
+  contentType: string;
+  sizeBytes: number;
+  filePath: string;
+}
+
+export interface ReportIssueEntry {
+  id: string;
+  sessionId: string;
+  timestamp: string;
+  title: string;
+  description: string;
+  category: ReportIssueCategory;
+  status: ReportIssueStatus;
+  attachmentIds: string[];
+  acknowledgedAt: string | null;
+  resolvedAt: string | null;
+  meta?: LogMeta;
+}
