+++ src/state/types/stats.ts
@@ -1,0 +1,32 @@
+export interface CoinStats {
+  one: number;
+  five: number;
+  ten: number;
+  twenty: number;
+}
+
+export interface JobStats {
+  total: number;
+  print: number;
+  copy: number;
+  scan: number;
+}
+
+export interface HopperSettings {
+  enabled: boolean;
+  timeoutMs: number;
+  retryCount: number;
+  dispenseCommandPrefix: string;
+  selfTestCommand: string;
+}
+
+export interface HopperStats {
+  dispenseAttempts: number;
+  dispenseSuccess: number;
+  dispenseFailures: number;
+  totalDispensed: number;
+  lastDispensedAt: string | null;
+  lastError: string | null;
+  selfTestPassed: boolean | null;
+  lastSelfTestAt: string | null;
+}
