+++ src/utils/lockout.ts
@@ -1,10 +1,11 @@
-import { db } from '@/services/db';
+import { stateRepository } from '@/state/repositories';
 
 export const MAX_ATTEMPTS = 3;
 export const LOCKOUT_DURATION_MS = 10 * 60 * 1000; // 10 minutes
 
 export function checkLockout(): { locked: boolean; remainingMs?: number } {
-  const { lockedUntil } = db.data!.adminLockout;
+  const state = stateRepository.getState();
+  const { lockedUntil } = state.adminLockout;
   if (!lockedUntil) return { locked: false };
 
   const expiry = new Date(lockedUntil).getTime();
@@ -15,30 +16,32 @@ export function checkLockout(): { locked: boolean; remainingMs?: number } {
   }
 
   // Lock expired — auto-clear
-  db.data!.adminLockout.failedAttempts = 0;
-  db.data!.adminLockout.lockedUntil = null;
-  db.write();
+  state.adminLockout.failedAttempts = 0;
+  state.adminLockout.lockedUntil = null;
+  void stateRepository.write();
   return { locked: false };
 }
 
 export async function recordFailedAttempt(): Promise<number> {
-  db.data!.adminLockout.failedAttempts += 1;
-  const attempts = db.data!.adminLockout.failedAttempts;
+  const state = stateRepository.getState();
+  state.adminLockout.failedAttempts += 1;
+  const attempts = state.adminLockout.failedAttempts;
 
   if (attempts >= MAX_ATTEMPTS) {
-    db.data!.adminLockout.lockedUntil = new Date(
+    state.adminLockout.lockedUntil = new Date(
       Date.now() + LOCKOUT_DURATION_MS,
     ).toISOString();
   }
 
-  await db.write();
+  await stateRepository.write();
   return attempts;
 }
 
 export async function clearLockout(): Promise<void> {
-  db.data!.adminLockout.failedAttempts = 0;
-  db.data!.adminLockout.lockedUntil = null;
-  await db.write();
+  const state = stateRepository.getState();
+  state.adminLockout.failedAttempts = 0;
+  state.adminLockout.lockedUntil = null;
+  await stateRepository.write();
 }
 
 export function formatRemainingTime(ms: number): string {
